+++ src/entities/link/api/hooks/index.ts
@@ -1,4 +1,4 @@
+export * from "./use-create";
 export * from "./use-fetch-by-id";
 export * from "./use-search-v1";
-export * from "./use-create";
 export * from "./use-update-link";
