+++ src/entities/link/api/hooks/use-create.ts
@@ -1,35 +1,16 @@
-import { useCallback, useContext } from "react";
-import { z } from "zod";
-import { LINK_SCHEMA } from "../../libs";
 import { ApiContext } from "@/shared/api";
+import { useCallback, useContext } from "react";
 import { CreateLinkRequestBody } from "../types";
 
 export const useCreateLink = () => {
   const api = useContext(ApiContext);
 
-  const createLink = useCallback(
-    async (formValues: z.infer<typeof LINK_SCHEMA>) => {
-      const body: CreateLinkRequestBody = {
-        forwardNameI18n: {
-          translations: {
-            en: formValues.name,
-          },
-        },
-        backwardNameI18n: {
-          translations: {
-            en: formValues.name,
-          },
-        },
-        ...formValues,
-      };
-
-      const { error } = await api.link.create({ body });
-      if (error) {
-        throw error;
-      }
-    },
-    []
-  );
+  const createLink = useCallback(async (body: CreateLinkRequestBody) => {
+    const { error } = await api.link.create({ body });
+    if (error) {
+      throw error;
+    }
+  }, []);
 
   return { createLink };
 };
