+++ src/entities/link/api/types.ts
@@ -33,7 +33,6 @@ export type LinkFilters = Partial<
   Pick<components["schemas"]["LinkSearchRqV1"], LinkFilterKeys>
 >;
 
-export type UpdateLinkRequestBody = components["schemas"]["LinkUpdateV1"];
 export type QueryLinkViewV1 = operations["linkViewV1"]["parameters"]["query"];
-
 export type CreateLinkRequestBody = components["schemas"]["LinkCreateV1"];
+export type UpdateLinkRequestBody = components["schemas"]["LinkUpdateV1"];
