+++ src/entities/link/index.ts
@@ -1,3 +1,3 @@
 export * from "./api";
-export * from "./libs";
 export * from "./components";
+export * from "./libs";
