+++ src/entities/link/libs/types.ts
@@ -1,4 +1,5 @@
 import { createEnum } from "@/shared/libs";
+import { z } from "zod";
 import { Link } from "../api";
 
 export type LinkType = NonNullable<Link["type"]>;
