--- src/entities/twin-class-link/components/resource-link/resource-link.tsx
@@ -1,36 +1,0 @@
-import { TwinClass, TwinClassContext } from "@/entities/twin-class";
-import { ResourceLink } from "@/shared/ui";
-import { Link2 } from "lucide-react";
-import { useContext } from "react";
-import { TwinClassLinkResourceTooltip } from "./tooltip";
-
-type Props = {
-  data: TwinClass;
-  disabled?: boolean;
-  withTooltip?: boolean;
-};
-//TODO LinkResourseLink rename
-//TODO move to entities link
-export function TwinClassLinkResourceLink({
-  data,
-  disabled,
-  withTooltip,
-}: Props) {
-  // const { twinClassId } = useContext(TwinClassContext);
-  const link = `/workspace/link/${data.id}`;
-
-  return (
-    <ResourceLink
-      IconComponent={() => <Link2 className="h-4 w-4" />}
-      data={data}
-      disabled={disabled}
-      renderTooltip={
-        withTooltip
-          ? (data) => <TwinClassLinkResourceTooltip data={data} link={link} />
-          : undefined
-      }
-      getDisplayName={(data) => data.name ?? ""}
-      link={link}
-    />
-  );
-}
