--- src/entities/twin-class-link/libs/helpers.ts
@@ -1,33 +1,0 @@
-import { TwinClass } from "@/entities/twin-class/api";
-import { RelatedObjects } from "@/shared/api";
-import { TwinClassLink, TwinClassLink_MANAGED } from "../api";
-
-export function hydrateLinkFromMap(
-  dto: TwinClassLink,
-  relatedObjects?: RelatedObjects
-): TwinClassLink_MANAGED {
-  const hydrated: TwinClassLink_MANAGED = Object.assign(
-    {},
-    dto
-  ) as TwinClassLink_MANAGED;
-
-  if (!relatedObjects?.twinClassMap) return hydrated;
-
-  if (dto.srcTwinClassId) {
-    hydrated.srcTwinClass = relatedObjects.twinClassMap[
-      dto.srcTwinClassId
-    ] as TwinClass;
-  }
-
-  if (dto.dstTwinClassId) {
-    hydrated.dstTwinClass = relatedObjects.twinClassMap[
-      dto.dstTwinClassId
-    ] as TwinClass;
-  }
-
-  if (dto.createdByUserId && relatedObjects.userMap) {
-    hydrated.createdByUser = relatedObjects.userMap[dto.createdByUserId];
-  }
-
-  return hydrated;
-}
