--- src/features/twinClass/index.ts
@@ -1,1 +1,0 @@
-export * from "./selectField";
