+++ src/widgets/form-fields/featurer/hooks/useSelectAdapter.ts
@@ -8,10 +8,9 @@ import {
   useTwinSelectAdapter,
   useTwinTouchIdSelectAdapter,
 } from "@/entities/twin";
+import { useTwinClassSelectAdapter } from "@/entities/twin-class";
 import { useTwinClassFieldSelectAdapter } from "@/entities/twin-class-field";
 import { useTwinStatusSelectAdapter } from "@/entities/twin-status";
-import { useTwinClassSelectAdapter } from "@/entities/twin-class";
-import { useLinkSelectAdapter } from "@/entities/twin-class-link";
 import { useTwinFlowSchemaSelectAdapter } from "@/entities/twinFlowSchema";
 import { useUserGroupSelectAdapter } from "@/entities/userGroup";
 import { createFixedSelectAdapter, SelectAdapter } from "@/shared/libs";
